refactor: 형변환 제거 및 InputParser를 view 패키지로 이동
