+++ src/test/java/blackjack/view/InputParserTest.java
@@ -1,4 +1,4 @@
-package blackjack.util;
+package blackjack.view;
 
 import static org.assertj.core.api.Assertions.assertThat;
